// backend/src/db/db.d.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export interface Customer {
  email: string | null;
  id: Generated<number>;
  name: string;
  phone: string | null;
}

export interface Permission {
  id: Generated<number>;
  key: string;
}

export interface Role {
  id: Generated<number>;
  name: string;
}

export interface RolePermission {
  permission_id: number;
  role_id: number;
}

export interface SchemaMigrations {
  version: string;
}

export interface User {
  id: Generated<number>;
  is_super_user: Generated<boolean>;
  password: string;
  refresh_token_version: Generated<number>;
  username: string;
}

export interface UserRole {
  role_id: number;
  user_id: number;
}

export interface DB {
  customer: Customer;
  permission: Permission;
  role: Role;
  role_permission: RolePermission;
  schema_migrations: SchemaMigrations;
  user: User;
  user_role: UserRole;
}